Conflicts:
	assessment/showtest.php
	course/course.php

+++ actions.php
@@ -8,7 +8,7 @@
 		$isgb = false;
 		$gb = '';
 	}
-	if(isset($_SERVER['HTTPS']) && $_SERVER['HTTPS']=='on') {
+	if((isset($_SERVER['HTTPS']) && $_SERVER['HTTPS']=='on') || (isset($_SERVER['HTTP_X_FORWARDED_PROTO']) && $_SERVER['HTTP_X_FORWARDED_PROTO']=='https'))  {
 		 $urlmode = 'https://';
 	} else {
 		 $urlmode = 'http://';
@@ -143,12 +143,19 @@
 						$error = 'Course is closed for self enrollment';
 					} else if ($_POST['ekey']=="" && $line['enrollkey'] != '') {
 						$error = 'No enrollment key provided';
-					} else if ($line['enrollkey'] != $_POST['ekey']) {
-						$error = 'Incorrect enrollment key';
 					} else {
-						$query = "INSERT INTO imas_students (userid,courseid) VALUES ('$newuserid','{$_POST['courseid']}');";
-						mysql_query($query) or die("Query failed : " . mysql_error());
-						echo '<p>You have been enrolled in course ID '.$_POST['courseid'].'</p>';
+						$keylist = array_map('trim',explode(';',$line['enrollkey']));
+						if (!in_array($_POST['ekey'], $keylist)) {
+							$error = 'Incorrect enrollment key';
+						} else {
+							if (count($keylist)>1) {
+								$query = "INSERT INTO imas_students (userid,courseid,section) VALUES ('$newuserid','{$_POST['courseid']}','{$_POST['ekey']}');";
+							} else {
+								$query = "INSERT INTO imas_students (userid,courseid) VALUES ('$newuserid','{$_POST['courseid']}');";
+							}
+							mysql_query($query) or die("Query failed : " . mysql_error());
+							echo '<p>You have been enrolled in course ID '.$_POST['courseid'].'</p>';
+						}
 					}
 				}
 				if ($error != '') {
@@ -302,12 +309,7 @@
 			echo "Please include Enrollment Key.  <a href=\"forms.php?action=enroll$gb\">Try Again</a>\n";
 			echo "</html></body>\n";
 			exit;
-		} else if ($line['enrollkey'] != $_POST['ekey']) {
-			echo "<html><body>\n";
-			echo "Incorrect Enrollment Key.  <a href=\"forms.php?action=enroll$gb\">Try Again</a>\n";
-			echo "</html></body>\n";
-			exit;
-		} else {
+		}  else {
 			$query = "SELECT * FROM imas_teachers WHERE userid='$userid' AND courseid='{$_POST['cid']}'";
 			$result = mysql_query($query) or die("Query failed : " . mysql_error());
 			if (mysql_num_rows($result)>0) {
@@ -335,8 +337,25 @@
 				echo "</html></body>\n";
 				exit;
 			} else {
-				$query = "INSERT INTO imas_students (userid,courseid) VALUES ('$userid','{$_POST['cid']}');";
-				mysql_query($query) or die("Query failed : " . mysql_error());
+				$keylist = array_map('trim',explode(';',$line['enrollkey']));
+				if (!in_array($_POST['ekey'], $keylist)) {
+					echo "<html><body>\n";
+					echo "Incorrect Enrollment Key.  <a href=\"forms.php?action=enroll$gb\">Try Again</a>\n";
+					echo "</html></body>\n";
+					exit;
+				} else {
+					if (count($keylist)>1) {
+						$query = "INSERT INTO imas_students (userid,courseid,section) VALUES ('$userid','{$_POST['cid']}','{$_POST['ekey']}');";
+					} else {
+						$query = "INSERT INTO imas_students (userid,courseid) VALUES ('$userid','{$_POST['cid']}');";
+					}
+					mysql_query($query) or die("Query failed : " . mysql_error());
+					echo '<p>You have been enrolled in course ID '.$_POST['courseid'].'</p>';
+				}
+				
+				
+				//$query = "INSERT INTO imas_students (userid,courseid) VALUES ('$userid','{$_POST['cid']}');";
+				//mysql_query($query) or die("Query failed : " . mysql_error());
 			}
 		}	
 	} else if ($_GET['action']=="unenroll") {
@@ -351,30 +370,38 @@
 			exit;
 		}
 		$cid = $_GET['cid'];
-		$query = "DELETE FROM imas_students WHERE userid='$userid' AND courseid='$cid'";
-		mysql_query($query) or die("Query failed : " . mysql_error());
-		$query = "SELECT id FROM imas_assessments WHERE courseid='$cid'";
+		$query = "SELECT allowunenroll FROM imas_courses WHERE id='$cid'";
 		$result = mysql_query($query) or die("Query failed : " . mysql_error());
-		while ($row = mysql_fetch_row($result)) {
-			$query = "DELETE FROM imas_assessment_sessions WHERE assessmentid='{$row[0]}' AND userid='$userid'";
+		if (mysql_result($result,0,0)==1) {
+			$query = "DELETE FROM imas_students WHERE userid='$userid' AND courseid='$cid'";
 			mysql_query($query) or die("Query failed : " . mysql_error());
-			$query = "DELETE FROM imas_exceptions WHERE assessmentid='{$row[0]}' AND userid='$userid'";
-			mysql_query($query) or die("Query failed : " . mysql_error());
-		}
-		$query = "SELECT id FROM imas_gbitems WHERE courseid='$cid'";
-		$result = mysql_query($query) or die("Query failed : " . mysql_error());
-		while ($row = mysql_fetch_row($result)) {
-			$query = "DELETE FROM imas_grades WHERE gradetype='offline' AND gradetypeid='{$row[0]}' AND userid='$userid'";
-			mysql_query($query) or die("Query failed : " . mysql_error());
-		}
-		$query = "SELECT id FROM imas_forums WHERE courseid='$cid'";
-		$result = mysql_query($query) or die("Query failed : " . mysql_error());
-		while ($row = mysql_fetch_row($result)) {
-			$q2 = "SELECT threadid FROM imas_forum_posts WHERE forumid='{$row[0]}'";
-			$r2 = mysql_query($q2) or die("Query failed : " . mysql_error());
-			while ($rw2 = mysql_fetch_row($r2)) {
-				$query = "DELETE FROM imas_forum_views WHERE threadid='{$rw2[0]}' AND userid='$userid'";
+			$query = "SELECT id FROM imas_assessments WHERE courseid='$cid'";
+			$result = mysql_query($query) or die("Query failed : " . mysql_error());
+			while ($row = mysql_fetch_row($result)) {
+				$query = "DELETE FROM imas_assessment_sessions WHERE assessmentid='{$row[0]}' AND userid='$userid'";
 				mysql_query($query) or die("Query failed : " . mysql_error());
+				$query = "DELETE FROM imas_exceptions WHERE assessmentid='{$row[0]}' AND userid='$userid'";
+				mysql_query($query) or die("Query failed : " . mysql_error());
+			}
+			
+			$query = "DELETE FROM imas_drillassess_sessions WHERE drillassessid IN (SELECT id FROM imas_drillassess WHERE courseid='$cid') AND userid='$userid'";
+			mysql_query($query) or die("Query failed : $query" . mysql_error());
+			
+			$query = "SELECT id FROM imas_gbitems WHERE courseid='$cid'";
+			$result = mysql_query($query) or die("Query failed : " . mysql_error());
+			while ($row = mysql_fetch_row($result)) {
+				$query = "DELETE FROM imas_grades WHERE gradetype='offline' AND gradetypeid='{$row[0]}' AND userid='$userid'";
+				mysql_query($query) or die("Query failed : " . mysql_error());
+			}
+			$query = "SELECT id FROM imas_forums WHERE courseid='$cid'";
+			$result = mysql_query($query) or die("Query failed : " . mysql_error());
+			while ($row = mysql_fetch_row($result)) {
+				$q2 = "SELECT threadid FROM imas_forum_posts WHERE forumid='{$row[0]}'";
+				$r2 = mysql_query($q2) or die("Query failed : " . mysql_error());
+				while ($rw2 = mysql_fetch_row($r2)) {
+					$query = "DELETE FROM imas_forum_views WHERE threadid='{$rw2[0]}' AND userid='$userid'";
+					mysql_query($query) or die("Query failed : " . mysql_error());
+				}
 			}
 		}
 	} else if ($_GET['action']=="chguserinfo") {
@@ -427,20 +454,21 @@
 		}
 				
 		$layoutstr = implode('|',$homelayout);
-		$query = "UPDATE imas_users SET FirstName='{$_POST['firstname']}',LastName='{$_POST['lastname']}',email='{$_POST['email']}',msgnotify=$msgnot,qrightsdef=$qrightsdef,deflib='$deflib',usedeflib='$usedeflib',homelayout='$layoutstr',listperpage='$perpage' ";
-		$query .= "WHERE id='$userid'";
-		mysql_query($query) or die("Query failed : " . mysql_error());
 		if (is_uploaded_file($_FILES['stupic']['tmp_name'])) {
 			processImage($_FILES['stupic'],$userid,200,200);
 			processImage($_FILES['stupic'],'sm'.$userid,40,40);
+			$chguserimg = ",hasuserimg=1";
 		} else if (isset($_POST['removepic'])) {
-			$curdir = rtrim(dirname(__FILE__), '/\\');
-			$galleryPath = "$curdir/course/files/";
-			if (file_exists($galleryPath.'userimg_'.$userid.'.jpg')) {
-				unlink($galleryPath.'userimg_'.$userid.'.jpg');
-				unlink($galleryPath.'userimg_sm'.$userid.'.jpg');
-			}
+			deletecoursefile('userimg_'.$userid.'.jpg');
+			deletecoursefile('userimg_sm'.$userid.'.jpg');
+			$chguserimg = ",hasuserimg=0";
+		} else {
+			$chguserimg = '';
 		}
+		$query = "UPDATE imas_users SET FirstName='{$_POST['firstname']}',LastName='{$_POST['lastname']}',email='{$_POST['email']}',msgnotify=$msgnot,qrightsdef=$qrightsdef,deflib='$deflib',usedeflib='$usedeflib',homelayout='$layoutstr',listperpage='$perpage'$chguserimg ";
+		$query .= "WHERE id='$userid'";
+		mysql_query($query) or die("Query failed : " . mysql_error());
+		
 		if (isset($_POST['dochgpw'])) {
 			$query = "SELECT password FROM imas_users WHERE id = '$userid'";
 			$result = mysql_query($query) or die("Query failed : " . mysql_error());
